Working script for updating post on Blogger but with flaw
The b.sh is a evolved combination of gen-blog-markdown.sh [1] and
gen-blog-rst.sh [2]. The script will have to work with patched GoogleCL [3].

The flaw is possibly in the server-side which adds unwanted <br/> to the
content. For example, a content like:

    <p>foobar</p>
    <p>text</p>

will end up like:

    <p>foobar</p><br/>
    <p>text</p>

There is a way to get around this, i.e. by eliminating newlines, however, if
the content is more complicated, such as with <script/> or <style/>. There is
no reliable way to get around. Therefore, using GData/GoogleCL isn't a viable
way to update or even to publish a blog post on Blogger.

The solution seems to be Blogger API V3 as mentioned in a discusstion [4].

[1] https://github.com/livibetter/dotfiles/blob/master/bin/gen-blog-markdown.sh
[2] https://github.com/livibetter/dotfiles/blob/master/bin/gen-blog-rst.sh
[3] https://gist.github.com/1196740#file-googlecl-blogger-update-r621-patch
[4] https://groups.google.com/d/topic/bloggerdev/qVNS8K7SF20/discussion